create copy of attributes when copying reader
