+++ pkg/stanza/fileconsumer/reader.go
@@ -128,7 +128,7 @@ func (r *Reader) curSplitFunc() bufio.SplitFunc {
 func (r *Reader) headerEmitFunc(ctx context.Context, attrs *FileAttributes, token []byte) {
 	if !r.headerSettings.matchRegex.Match(token) {
 		r.HeaderFinalized = true
-		attrs.HeaderAttributes = r.HeaderAttributes
+		attrs.headerAttributes = r.HeaderAttributes
 
 		if err := r.headerPipeline.Stop(); err != nil {
 			r.SugaredLogger.Errorw("Failed to stop header pipeline during finalization", zap.Error(err))
@@ -206,4 +206,20 @@ func min0(a, b int) int {
 		return a
 	}
 	return b
+}
+
+// mapCopy deep copies the provided attributes map.
+func mapCopy(m map[string]any) map[string]any {
+	newMap := make(map[string]any, len(m))
+	for k, v := range m {
+		switch typedVal := v.(type) {
+		case map[string]any:
+			newMap[k] = mapCopy(typedVal)
+		default:
+			// Assume any other values are safe to directly copy.
+			// Struct types and slice types shouldn't appear in attribute maps from pipelines
+			newMap[k] = v
+		}
+	}
+	return newMap
 }
